+++ packages/passport-crypto/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./endToEndEncryption";
 export * from "./passportCrypto";
 export * from "./types";
+export * from "./utils";
