+++ packages/passport-crypto/src/passportCrypto.ts
@@ -53,7 +53,13 @@ export class PCDCrypto {
       length,
       utils.stringToArrayBuffer(password),
       utils.hexStringToArrayBuffer(salt),
+      // `oplimit` represents the maximum amount of computations to perform
+      // for generating a key. `crypto_pwhash_OPSLIMIT_INTERACTIVE` is recommended
+      // for intereactive, online applications such as PCDPass.
+      // Source: https://libsodium.gitbook.io/doc/password_hashing/default_phf
       this.sodium.crypto_pwhash_OPSLIMIT_INTERACTIVE,
+      // `memlimit` represents the maximum amount of RAM in bytes thet function
+      // will use. It is required to use the same values for `opslimit` and `memlimit`.
       this.sodium.crypto_pwhash_MEMLIMIT_INTERACTIVE,
       this.sodium.crypto_pwhash_ALG_DEFAULT,
       "hex"
